remove create function, pour out the hooks into the hooks folder
